requirements.txt
swagger-ui-bundle==0.0.9
connexion==2.14.1
sqlalchemy
base
mysql-connector-python
pymysql==1.0.2
logger
logging.config
uuid
mysql.connector==2.2.9